refactor(eternal-jukebox): Rename components
