+++ custom-apps/eternal-jukebox/src/components/visualizer/JukeboxVisualizer.module.scss
@@ -1,0 +1,4 @@
+.jukebox-graph {
+    width: 100%;
+    height: 100%;
+}
