+++ custom-apps/eternal-jukebox/src/components/visualizer/VisualizerEdge.module.scss
@@ -1,0 +1,7 @@
+.edge-path {
+    stroke-opacity: 0.7;
+
+    &:hover {
+        stroke-opacity: 1;
+    }
+}
