+++ custom-apps/eternal-jukebox/src/css/app.module.scss
@@ -4,39 +4,6 @@
     align-items: center;
 }
 
-.jukebox-header {
-    display: flex;
-    justify-content: center;
-    align-items: stretch;
-
-    width: 100%;
-
-    .side {
-        flex-grow: 1;
-
-        display: flex;
-        align-items: start;
-        justify-content: end;
-
-        padding-left: 2rem;
-        padding-right: 2rem;
-    }
-
-    .center {
-        * {
-            text-align: center;
-        }
-    }
-}
-
-.container {
-    display: flex;
-    justify-content: start;
-    align-items: center;
-    flex-direction: column;
-    height: 100%;
-}
-
 .full-size-container {
     position: absolute;
     left: 0;
@@ -72,39 +39,4 @@
             flex-direction: column;
         }
     }
-}
-
-.visualizer-container {
-    flex-grow: 1;
-}
-
-.stats {
-    width: 100%;
-    display: flex;
-    justify-content: space-around;
-    padding-top: 1rem;
-    padding-bottom: 1rem;
-
-    span {
-        flex-grow: 1;
-        flex-basis: 0;
-        text-align: center;
-    }
-}
-
-.jukebox-graph {
-    width: 100%;
-    height: 100%;
-
-    .edge-path {
-        stroke-opacity: 0.7;
-
-        &:hover {
-            stroke-opacity: 1;
-        }
-    }
-}
-
-.small-text {
-    font-size: small;
 }
