+++ custom-apps/eternal-jukebox/tsconfig.json
@@ -1,7 +1,7 @@
 {
     "extends": "../../tsconfig.base.json",
     "compilerOptions": {
-        "outDir": "./dist",
+        "outDir": "./dist"
     },
     "include": ["./src/**/*", "../../libs/shared/src/types/**/*"]
 }
