+++ libs/shared/src/components/inputs/MultiRangeSlider/MultiRangeSlider.module.scss
