+++ libs/shared/src/components/settings/CheckboxContainer/CheckboxContainer.module.scss
