+++ libs/shared/src/components/settings/SliderContainer/SliderContainer.module.scss
